Removal of the unused svn:mime-type subversion property from all of FGAddon.
This follows from the proposal for deleting the property for all XML files at
http://thread.gmane.org/gmane.games.flightgear.devel/77365 and the discussions
in the thread at http://thread.gmane.org/gmane.games.flightgear.devel/77606 (for
setting up a repository pre-commit hook script for catching this unused
property) where deleting the property for all files was discussed.

This svn:mime-type repository property is only for optimising the back-end
storage size of the db/revs/* files on the SourceForge infrastructure, by
allowing certain text files with a very limited number of extremely long lines
to be treated as a binary rather than a text file, hence suppressing diffs in
the commit.  As no one has deliberately set this and many text files in FGAddon
are accidentally not being treated as text, this change will result in a major
repository clean up which will allow all text files to be treated as text.  The
activation of the pre-commit hook will prevent this problem from ever
reappearing.

This change touches the files in 480 aircraft directories out of 485, however
this does not affect the file contents or aircraft in any way.



git-svn-id: https://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/T-50@583 9eb9bc84-db9d-4e05-893b-26dd41ac94a6